changed controller using one controller to support services added class encapsulation for DynamixelBus and DynamixelDevice using Timer for publishers
